Move FastValueView to separate files.
