+++ eval/src/vespa/eval/eval/fast_value.hpp
@@ -1,7 +1,7 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "value_builder_factory.h"
-#include "fast_addr_map.h"
+#include "fast_value_index.h"
 #include "inline_operation.h"
 #include <vespa/eval/instruction/generic_join.h>
 #include <vespa/vespalib/stllike/hashtable.hpp>
@@ -11,16 +11,6 @@
 namespace vespalib::eval {
 
 //-----------------------------------------------------------------------------
-
-// This is the class instructions will look for when optimizing sparse
-// operations by calling inline functions directly.
-struct FastValueIndex final : Value::Index {
-    FastAddrMap map;
-    FastValueIndex(size_t num_mapped_dims_in, const StringIdVector &labels, size_t expected_subspaces_in)
-        : map(num_mapped_dims_in, labels, expected_subspaces_in) {}
-    size_t size() const override { return map.size(); }
-    std::unique_ptr<View> create_view(ConstArrayRef<size_t> dims) const override;
-};
 
 inline bool is_fast(const Value::Index &index) {
     return (std::type_index(typeid(index)) == std::type_index(typeid(FastValueIndex)));
